requirements.txt
opencv-python
torchtext==0.10.0
torch==1.9.0
transformers==0.4
tqdm
pandas
Pillow
pmcore@git+https://bff59f9534c6b885295b0709332e6e393f1593b3@github.com/AISLE-3/product-matching-core